shared/requirements.txt
# Shared Library Requirements
pydantic>=2.0.0
pyyaml>=6.0
pillow>=9.0.0

# HEIC/HEIF Support
pillow-heif>=0.13.0

# Testing
pytest>=7.4.0